add new options to remove old logs when there is not enough space
new options:
  --purge_free_space_threshold=<total_space_in_MB>
  --uncompressed_keep=<hours>
  --uncompressed_free_space_threshold=<total_space_in_MB>

I added the condition when there is less then --purge_free_space_threshold=xxxMB space available. Then the oldest
archives are being removed until there is not enough space. The default threshold is 200MB.

The features for uncompressed logs add the option to specify the number of hours worth of uncompressed logs
to keep around. It's similar to --purge option but the old uncompressed logs are removed only when there is
not enough space available. The default threshold for disk space is 100MB and it's kept 24 hours of logs.

Signed-off-by: Pavel Pulec <kayn@inuits.eu>